use ULONG instead of unsigned long
Define it as uint32_t though.

+++ include/stdint.h
@@ -32,8 +32,8 @@ typedef signed int int_least32_t;
 typedef unsigned int uint_least32_t;
 typedef signed int int_fast32_t;
 typedef unsigned int uint_fast32_t;
-typedef signed long int int32_t;
-typedef unsigned long int uint32_t;
+typedef signed int int32_t;
+typedef unsigned int uint32_t;
 
 __extension__ typedef signed long long int int_least64_t;
 __extension__ typedef unsigned long long int uint_least64_t;
